pyproject.toml
[project]
name = "drs4"
version = "0.1.0"
description = "Control and data acquisition software for FINER/DRS4"
readme = "README.md"
keywords = [
    "astronomy",
    "drs4",
    "finer",
    "lmt",
    "millimeter",
    "python",
    "radio-astronomy",
    "single-dish",
    "submillimeter",
]
requires-python = ">=3.10,<3.15"
dependencies = [
    "dateparser>=1,<2",
    "fire>=0.7,<1",
    "matplotlib>=3,<4",
    "numpy>=2,<3",
    "pillow>=10,<11",
    "tqdm>=4,<5",
    "typing-extensions>=4,<5",
    "xarray>=2024,<2027",
    "xarray-dataclasses>=1,<2",
    "zarr>=2,<4",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[[project.authors]]
name = "Akio Taniguchi"
email = "a-taniguchi@mail.kitami-it.ac.jp"

[project.license]
file = "LICENSE"

[project.urls]
homepage = "https://finerreceiver.github.io/drs4"
repository = "https://github.com/finerreceiver/drs4"

[dependency-groups]
dev = [
    "black>=25,<27",
    "ipython>=8,<10",
    "myst-parser>=3,<5",
    "pandas-stubs>=2,<3",
    "pydata-sphinx-theme>=0.16,<1.0",
    "pyright>=1,<2",
    "pytest>=9,<10",
    "sphinx>=8,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
typeCheckingMode = "basic"
reportIncompatibleMethodOverride = "warning"
reportUnknownMemberType = "warning"
